wsd: move WopiStorage to own home
