+++ Makefile.am
@@ -151,7 +151,8 @@ coolwsd_sources = common/Crypto.cpp \
                   wsd/QuarantineUtil.cpp \
                   wsd/wopi/CheckFileInfo.cpp \
                   wsd/wopi/StorageConnectionManager.cpp \
-                  wsd/wopi/WopiProxy.cpp
+                  wsd/wopi/WopiProxy.cpp \
+                  wsd/wopi/WopiStorage.cpp
 
 coolwsd_SOURCES = $(coolwsd_sources) \
                   $(shared_sources) \
@@ -314,7 +315,8 @@ wsd_headers = wsd/Admin.hpp \
               wsd/HostUtil.hpp \
               wsd/wopi/CheckFileInfo.hpp \
               wsd/wopi/StorageConnectionManager.hpp \
-              wsd/wopi/WopiProxy.hpp
+              wsd/wopi/WopiProxy.hpp \
+              wsd/wopi/WopiStorage.hpp
 
 shared_headers = common/Common.hpp \
                  common/Clipboard.hpp \
